Dirty Fix the Cookie issue within SystemWebCookieManager
