+++ myWebApp/Site.Master
@@ -53,6 +53,7 @@
                         <li><a runat="server" href="~/">Home</a></li>
                         <li><a runat="server" href="~/About">About</a></li>
                         <li><a runat="server" href="~/Contact">Contact</a></li>
+                        <li><a runat="server" href="~/Contact">Hi, <asp:Label ID="lblUsername" runat="server" Text=""></asp:Label></a></li>
                     </ul>
                 </div>
             </div>
