+++ myWebApp/Site.Master.cs
@@ -5,13 +5,14 @@ using System.Web;
 using System.Web.UI;
 using System.Web.UI.WebControls;
 
+
 namespace myWebApp
 {
     public partial class SiteMaster : MasterPage
     {
         protected void Page_Load(object sender, EventArgs e)
         {
-
+            lblUsername.Text = HttpContext.Current.User.Identity.Name;
         }
     }
 }
