+++ myWebApp/Web.config
@@ -19,7 +19,7 @@
 		<add key="DbConnectionString2" value="Server=localhost;Database=AdventureWorks2019;Trusted_Connection=false; User Id=webappuser; Password=Welcome123; min pool size=3; Application Name=WebApp;" />
 		<add key="DbConnectionString1" value="Server=localhost;Database=StackOverflow2013;Trusted_Connection=false; User Id=webappuser; Password=Welcome123; min pool size=3; Application Name=WebApp;" />
 		<add key="DbConnectionString" value="Data Source=localhost\MSSQLDEV2016; Initial Catalog=PeopleWeb_Logbook; User Id=PeopleWeb; Password=Welcome123; Max Pool Size=1; Connection Timeout=25;Application Name=MyWebApp" />
-		<add key="CacheConnectionString" value="51.145.88.178:6666,abortConnect=false" />
+		<add key="CacheConnectionString" value="20.49.207.78:6666,abortConnect=false" />
 		<add key="auth0:Domain" value="dev-sso-poc.auth0.com" />
 		<add key="auth0:ClientId" value="CRjvhgle3aqP263DIeY0RTvaPuesnlRi" />
 		<add key="auth0:RedirectUri" value="https://localhost:44347/pages/login" />
@@ -27,7 +27,7 @@
 	</appSettings>
 	<log4net>
 		<appender name="LogFileAppender" type="log4net.Appender.RollingFileAppender">
-			<param name="File" value="D:\\Temp\\myWebApp.log" />
+			<param name="File" value="myWebApp.log" />
 			<lockingModel type="log4net.Appender.FileAppender+MinimalLock" />
 			<appendToFile value="true" />
 			<rollingStyle value="Size" />
@@ -123,6 +123,10 @@
 			<dependentAssembly>
 				<assemblyIdentity name="Microsoft.Extensions.Primitives" publicKeyToken="adb9793829ddae60" culture="neutral" />
 				<bindingRedirect oldVersion="0.0.0.0-5.0.0.0" newVersion="5.0.0.0" />
+			</dependentAssembly>
+			<dependentAssembly>
+				<assemblyIdentity name="Microsoft.Owin.Security.OAuth" publicKeyToken="31bf3856ad364e35" culture="neutral" />
+				<bindingRedirect oldVersion="0.0.0.0-4.2.2.0" newVersion="4.2.2.0" />
 			</dependentAssembly>
 		</assemblyBinding>
 	</runtime>
