+++ myWebApp/packages.config
@@ -64,7 +64,7 @@
   <package id="Microsoft.Owin.Host.SystemWeb" version="4.2.2" targetFramework="net472" />
   <package id="Microsoft.Owin.Security" version="4.2.2" targetFramework="net472" />
   <package id="Microsoft.Owin.Security.Cookies" version="4.2.2" targetFramework="net472" />
-  <package id="Microsoft.Owin.Security.OAuth" version="3.0.1" targetFramework="net472" />
+  <package id="Microsoft.Owin.Security.OAuth" version="4.2.2" targetFramework="net472" />
   <package id="Microsoft.Owin.Security.OpenIdConnect" version="4.2.2" targetFramework="net472" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net472" />
   <package id="Modernizr" version="2.8.3" targetFramework="net472" />
